Fifth Commit - Exception Handling

+++ src/main/java/com/enviro/assessment/grad001/lutho/nondala/controller/CategoryController.java
@@ -2,13 +2,19 @@ package com.enviro.assessment.grad001.lutho.nondala.controller;
 
 import com.enviro.assessment.grad001.lutho.nondala.entity.Category;
 import com.enviro.assessment.grad001.lutho.nondala.service.CategoryService;
+import jakarta.validation.Valid;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.FieldError;
+import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 
 @RestController
 @RequestMapping("enviro/category/")
@@ -18,8 +24,11 @@ public class CategoryController {
     private CategoryService service;
 
     @PostMapping("create")
-    public ResponseEntity<Category> create(@RequestBody Category category){
+    public ResponseEntity<Category> create(@Valid @RequestBody Category category, BindingResult result){
         try {
+            if (result.hasErrors()){
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            }
             return ResponseEntity.status(HttpStatus.OK).body(this.service.create(category));
         } catch (Exception e){
             System.out.println(e.getMessage());
@@ -38,8 +47,11 @@ public class CategoryController {
     }
 
     @PutMapping("update")
-    public ResponseEntity<Category> update(@RequestBody Category category){
+    public ResponseEntity<Category> update(@Valid @RequestBody Category category, BindingResult result){
         try {
+            if (result.hasErrors() || category.getId() == 0 || String.valueOf(category.getId()) == null){
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            }
             return ResponseEntity.status(HttpStatus.OK).body(this.service.update(category));
         } catch (Exception e){
             System.out.println(e.getMessage());
@@ -48,11 +60,16 @@ public class CategoryController {
     }
 
     @DeleteMapping("delete/{id}")
-    public void delete(@PathVariable long id){
+    public ResponseEntity<?> delete(@PathVariable long id){
         try {
+            if (id == 0 || String.valueOf(id) == null){
+                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(null);
+            }
             this.service.delete(id);
+            return ResponseEntity.status(HttpStatus.OK).body("Delete Successful!");
         } catch (Exception e){
             System.out.println(e.getMessage());
+            return null;
         }
     }
 
@@ -65,4 +82,5 @@ public class CategoryController {
             return null;
         }
     }
+
 }
